Make build populate docs dir from submodule
This change adds https://github.com/relaxng/relaxng.org as a submodule
and updates the (ant) build to make it populate the `docs` subdirectory
by copying over files from the (new) relaxng.org submodule.

The build needs the `docs` content in order to create the jing and trang
zip release distributions with the docs included.